telegram_bot/app/requirements.txt
python-telegram-bot==21.3
openai==1.35.3
logfmter==0.0.8
asyncpg==0.29.0
telegramify-markdown==0.1.8
